+++ plugin.ts
@@ -2,7 +2,7 @@
  * @name raider-io
  * @pluginURL https://raw.githubusercontent.com/rileyio/raider-io/master/plugin.ts
  * @repo rileyio/raider-io
- * @version 1.1.1
+ * @version 1.2.0
  */
 
 import * as moment from 'moment'
@@ -14,6 +14,7 @@ import { fetchSeasonCutoffs, getMythicPlusScorePlacement } from './mythic-plus'
 import { CharacterProfile } from './character'
 import { Plugin } from '../../src/objects/plugin'
 import axios from 'axios'
+import { realms } from './servers'
 
 // const Covenant = {
 //   Kyrian: '<:Kyrian:1008063029830746254>',
@@ -30,49 +31,55 @@ const GroupRole = {
 
 export class RaiderIOPlugin extends Plugin {
   config = { baseURL: 'https://raider.io/api/v1' }
+  routes = [
+    new RouteConfiguration({
+      autocomplete: {
+        options: {
+          realm: realms
+        }
+      },
+      category: 'Plugin/Raider.IO',
+      controller: this.routeCommand,
+      name: 'rio',
+      permissions: {
+        defaultEnabled: false,
+        serverOnly: false
+      },
+      plugin: this,
+      slash: new SlashCommandBuilder()
+        .setName('rio')
+        .setDescription('Raider.io')
+        // * Character Profile
+        .addSubcommand((subcommand) =>
+          subcommand
+            .setName('character-profile')
+            .setDescription('Lookup Gear, Guild, Covenant, M+ and Raid Progression')
+            .addStringOption((option) =>
+              option
+                .setName('region')
+                .setDescription('Region')
+                .setRequired(true)
+                .setChoices({ name: 'US', value: 'us' }, { name: 'EU', value: 'eu' }, { name: 'TW', value: 'tw' }, { name: 'KR', value: 'kr' }, { name: 'CN', value: 'cn' })
+            )
+            .addStringOption((option) => option.setName('realm').setDescription('Server/Realm (Example: Area 52)').setRequired(true).setAutocomplete(true))
+            .addStringOption((option) => option.setName('name').setDescription('Character Name').setRequired(true))
+        ),
+      type: 'discord-chat-interaction'
+    })
+  ]
 
   constructor() {
     super()
     this.autoCheckForUpdate = false
   }
 
-  public async onEnabled() {
-    await this.bot.Router.addRoute(
-      new RouteConfiguration({
-        category: 'Plugin/Raider.IO',
-        controller: this.routeCommand,
-        name: 'rio',
-        permissions: {
-          defaultEnabled: false,
-          serverOnly: false
-        },
-        plugin: this,
-        slash: new SlashCommandBuilder()
-          .setName('rio')
-          .setDescription('Raider.io')
-          // * Character Profile
-          .addSubcommand((subcommand) =>
-            subcommand
-              .setName('character-profile')
-              .setDescription('Lookup Gear, Guild, Covenant, M+ and Raid Progression')
-              .addStringOption((option) =>
-                option
-                  .setName('region')
-                  .setDescription('Region')
-                  .setRequired(true)
-                  .setChoices({ name: 'US', value: 'us' }, { name: 'EU', value: 'eu' }, { name: 'TW', value: 'tw' }, { name: 'KR', value: 'kr' }, { name: 'CN', value: 'cn' })
-              )
-              .addStringOption((option) => option.setName('realm').setDescription('Server/Realm (Example: area-52)').setRequired(true))
-              .addStringOption((option) => option.setName('name').setDescription('Character Name').setRequired(true))
-          ),
-        type: 'discord-chat-interaction'
-      })
-    )
-  }
+  // onEnabled = async () => {
+  //   console.log('test')
+  // }
 
-  public async onDisabled() {
-    await this.bot.Router.removeRoute('rio')
-  }
+  // onDisabled = async () => {
+  //   await this.bot.Router.removeRoute('rio')
+  // }
 
   public async fetchCharacterProfile(plugin: RaiderIOPlugin, routed: Routed<'discord-chat-interaction'>) {
     const region = routed.interaction.options.get('region')?.value as string
@@ -84,7 +91,7 @@ export class RaiderIOPlugin extends Plugin {
     charURL += 'covenant,'
     charURL += 'raid_progression,'
     charURL += 'raid_achievement_curve:castle-nathria:sanctum-of-domination:sepulcher-of-the-first-ones:vault-of-the-incarnates,'
-    charURL += 'mythic_plus_scores_by_season:season-df-1:season-sl-4:season-sl-3:season-sl-2:season-sl-1'
+    charURL += 'mythic_plus_scores_by_season:season-df-2:season-df-1:season-sl-4:season-sl-3:season-sl-2:season-sl-1'
 
     console.log('Character URL:', encodeURI(charURL))
 
@@ -143,8 +150,10 @@ export class RaiderIOPlugin extends Plugin {
       const sls3 = data.mythic_plus_scores_by_season.find((s) => s.season === 'season-sl-3')
       const sls4 = data.mythic_plus_scores_by_season.find((s) => s.season === 'season-sl-4')
       const dfs1 = data.mythic_plus_scores_by_season.find((s) => s.season === 'season-df-1')
+      const dfs2 = data.mythic_plus_scores_by_season.find((s) => s.season === 'season-df-2')
       const seasonCutoffs = await fetchSeasonCutoffs(this.config.baseURL, region)
-      const dfs1MythicPlusScorePlacement = seasonCutoffs && dfs1 ? getMythicPlusScorePlacement(seasonCutoffs, dfs1.scores.all) : null
+      const dfs1MythicPlusScorePlacement = undefined //seasonCutoffs && dfs1 ? getMythicPlusScorePlacement(seasonCutoffs, dfs1.scores.all) : null
+      const dfs2MythicPlusScorePlacement = seasonCutoffs && dfs2 ? getMythicPlusScorePlacement(seasonCutoffs, dfs2.scores.all) : null
 
       // TODO: Create cache to not spam these old seasons constantly
       const sls4MythicPlusScorePlacement = false
@@ -153,10 +162,18 @@ export class RaiderIOPlugin extends Plugin {
       //const sls3MythicPlusScorePlacement = seasonCutoffs && sls3 ? getMythicPlusScorePlacement(seasonCutoffs, sls3.scores.all) : null
 
       // Dragonflight Scores
-      if (dfs1) {
-        if (dfs1MythicPlusScorePlacement) mPlus += `**DF S1** \`${dfs1.scores.all}\` | **${dfs1MythicPlusScorePlacement}**\n`
-        else mPlus += `**DF S1** \`${dfs1.scores.all}\` | **${dfs1MythicPlusScorePlacement}**\n`
-        // Role Scores
+      if (dfs2.scores.all) {
+        if (dfs2MythicPlusScorePlacement) mPlus += `**DF S2** \`${dfs2.scores.all}\`${dfs2MythicPlusScorePlacement ? ` | **${dfs2MythicPlusScorePlacement}**` : ''}\n`
+        else mPlus += `**DF S2** \`${dfs2.scores.all}\`\n`
+        // Role Scores (s2)
+        if (dfs2.scores.tank) mPlus += `${GroupRole.Tank} \`${dfs2.scores.tank}\``
+        if (dfs2.scores.healer) mPlus += ` ${GroupRole.Healer} \`${dfs2.scores.healer}\``
+        if (dfs2.scores.dps) mPlus += ` ${GroupRole.DPS} \`${dfs2.scores.dps}\``
+      }
+      if (dfs1.scores.all) {
+        if (dfs1MythicPlusScorePlacement) mPlus += `**DF S1** \`${dfs1.scores.all}\`${dfs1MythicPlusScorePlacement ? ` | **${dfs1MythicPlusScorePlacement}**` : ''}\n`
+        else mPlus += `\n\n**DF S1** \`${dfs1.scores.all}\`\n`
+        // Role Scores (s1)
         if (dfs1.scores.tank) mPlus += `${GroupRole.Tank} \`${dfs1.scores.tank}\``
         if (dfs1.scores.healer) mPlus += ` ${GroupRole.Healer} \`${dfs1.scores.healer}\``
         if (dfs1.scores.dps) mPlus += ` ${GroupRole.DPS} \`${dfs1.scores.dps}\``
@@ -165,7 +182,7 @@ export class RaiderIOPlugin extends Plugin {
       // Shadowlands Scores
       if (sls1 || sls2 || sls3 || sls4) mPlus += `\n\n**==== Shadowlands ====**`
 
-      if (sls4) {
+      if (sls4.scores.all) {
         if (sls4MythicPlusScorePlacement) mPlus += `\n\n**SL S4** \`${sls4.scores.all}\ | **${sls4MythicPlusScorePlacement}**\n`
         else mPlus += `\n\n**SL S4** \`${sls4.scores.all}\`\n`
         // Role Scores
@@ -173,7 +190,7 @@ export class RaiderIOPlugin extends Plugin {
         if (sls4.scores.healer) mPlus += ` ${GroupRole.Healer} \`${sls4.scores.healer}\``
         if (sls4.scores.dps) mPlus += ` ${GroupRole.DPS} \`${sls4.scores.dps}\``
       }
-      if (sls3) {
+      if (sls3.scores.all) {
         if (sls3MythicPlusScorePlacement) mPlus += `\n\n**SL S3** \`${sls3.scores.all}\` | **${sls3MythicPlusScorePlacement}**\n`
         else mPlus += `\n\n**SL S3** \`${sls3.scores.all}\`\n`
         // Role Scores
@@ -181,20 +198,20 @@ export class RaiderIOPlugin extends Plugin {
         if (sls3.scores.healer) mPlus += ` ${GroupRole.Healer} \`${sls3.scores.healer}\``
         if (sls3.scores.dps) mPlus += ` ${GroupRole.DPS} \`${sls3.scores.dps}\``
       }
-      if (sls2) {
-        mPlus += `\n\n**SL S2** \`${sls2.scores.all}\`\n`
-        // Role Scores
-        if (sls2.scores.tank) mPlus += `${GroupRole.Tank} \`${sls2.scores.tank}\``
-        if (sls2.scores.healer) mPlus += ` ${GroupRole.Healer} \`${sls2.scores.healer}\``
-        if (sls2.scores.dps) mPlus += ` ${GroupRole.DPS} \`${sls2.scores.dps}\``
-      }
-      if (sls1) {
-        mPlus += `\n\n**SL S1** \`${sls1.scores.all}\`\n`
-        // Role Scores
-        if (sls1.scores.tank) mPlus += `${GroupRole.Tank} \`${sls1.scores.tank}\``
-        if (sls1.scores.healer) mPlus += ` ${GroupRole.Healer} \`${sls1.scores.healer}\``
-        if (sls1.scores.dps) mPlus += ` ${GroupRole.DPS} \`${sls1.scores.dps}\``
-      }
+      // if (sls2.scores.all) {
+      //   mPlus += `\n\n**SL S2** \`${sls2.scores.all}\`\n`
+      //   // Role Scores
+      //   if (sls2.scores.tank) mPlus += `${GroupRole.Tank} \`${sls2.scores.tank}\``
+      //   if (sls2.scores.healer) mPlus += ` ${GroupRole.Healer} \`${sls2.scores.healer}\``
+      //   if (sls2.scores.dps) mPlus += ` ${GroupRole.DPS} \`${sls2.scores.dps}\``
+      // }
+      // if (sls1.scores.all) {
+      //   mPlus += `\n\n**SL S1** \`${sls1.scores.all}\`\n`
+      //   // Role Scores
+      //   if (sls1.scores.tank) mPlus += `${GroupRole.Tank} \`${sls1.scores.tank}\``
+      //   if (sls1.scores.healer) mPlus += ` ${GroupRole.Healer} \`${sls1.scores.healer}\``
+      //   if (sls1.scores.dps) mPlus += ` ${GroupRole.DPS} \`${sls1.scores.dps}\``
+      // }
 
       // Page M+ Footer
       mPlus += '\n\n'
@@ -206,12 +223,21 @@ export class RaiderIOPlugin extends Plugin {
       // if (data.mythic_plus_scores.healer) description += `\n- Healer \`${data.mythic_plus_scores.healer}\``
 
       let raid = ''
+      const atsc = data.raid_progression['aberrus-the-shadowed-crucible']
       const voi = data.raid_progression['vault-of-the-incarnates']
 
       // const curveCN = data.raid_achievement_curve.find((r) => r.raid === 'castle-nathria')
       // const curveSoD = data.raid_achievement_curve.find((r) => r.raid === 'sanctum-of-domination')
       // const curveSoFO = data.raid_achievement_curve.find((r) => r.raid === 'sepulcher-of-the-first-ones')
+      const curveATSC = data.raid_achievement_curve.find((r) => r.raid === 'aberrus-the-shadowed-crucible')
       const curveVoI = data.raid_achievement_curve.find((r) => r.raid === 'vault-of-the-incarnates')
+
+      if (atsc) {
+        raid += `\n\n**Aberrus the Shadowed Crucible** ${curveATSC ? (curveATSC.aotc ? '`[AOTC]`' : '') : ''} ${curveATSC ? (curveATSC.cutting_edge ? '`[CE]`' : '') : ''}`
+        raid += `\nMythic ${atsc.mythic_bosses_killed}/${atsc.total_bosses} ${curveATSC?.cutting_edge ? '`' + (curveATSC?.cutting_edge).substring(0, 10) + '`' : ''}`
+        raid += `\nHeroic ${atsc.heroic_bosses_killed}/${atsc.total_bosses} ${curveATSC?.aotc ? '`' + (curveATSC?.aotc).substring(0, 10) + '`' : ''}`
+        raid += `\nNormal ${atsc.normal_bosses_killed}/${atsc.total_bosses}`
+      }
 
       if (voi) {
         raid += `\n\n**Vault of the Incarnates** ${curveVoI ? (curveVoI.aotc ? '`[AOTC]`' : '') : ''} ${curveVoI ? (curveVoI.cutting_edge ? '`[CE]`' : '') : ''}`
@@ -239,7 +265,10 @@ export class RaiderIOPlugin extends Plugin {
       // The public message posted
       let publicMsg: Message<boolean>
 
-      const collector = routed.channel?.createMessageComponentCollector({ time: 5 * (60 * 1000) })
+      const collector = routed.channel?.createMessageComponentCollector({
+        filter: (i) => i.user.id === routed.interaction.user.id && i.message.interaction.id === routed.interaction.id,
+        time: 5 * (60 * 1000)
+      })
       collector?.on('collect', async (i: ButtonInteraction) => {
         // Enable Make Public button now
         completeOptions.components[0].setDisabled(false)
@@ -278,10 +307,18 @@ export class RaiderIOPlugin extends Plugin {
 
         // When stopped
         if (i.customId === 'end') {
-          await i.update({
-            components: [],
-            embeds: publicMsg ? [] : [lastMsg]
-          })
+          // At least 1 button has been pressed
+          if (lastMsg)
+            await i.update({
+              components: [],
+              embeds: publicMsg ? [] : [lastMsg]
+            })
+          // Condition for when no other buttons are pressed
+          else
+            await i.update({
+              components: [],
+              content: 'Done'
+            })
           collector.stop('stopped')
         }
 
@@ -297,9 +334,11 @@ export class RaiderIOPlugin extends Plugin {
 
       collector?.on('end', async (collected, reason) => {
         if (reason && reason !== 'stopped') {
-          routed.interaction.editReply({
+          // If there is an embed, update it with the last message
+          await routed.interaction.editReply({
             components: [],
-            embeds: [lastMsg]
+            content: 'Looks like you took too long to respond.',
+            embeds: lastMsg ? [lastMsg] : undefined
           })
         }
       })
@@ -308,7 +347,7 @@ export class RaiderIOPlugin extends Plugin {
       completeOptions.components[0].setDisabled(true)
 
       // First post with button options
-      await routed.interaction.editReply({
+      return await routed.interaction.editReply({
         components
       })
     } catch (error) {