+++ CHANGELOG.md
@@ -1,14 +1,39 @@
 # Change Log
 
-## 1.1.0
+## 2.0.0
+### 🛠️ Fixes
+- Fixed: Issue with replying to multiple pending requests.
+- Fixed: Lookup but no interaction timeout error.
+- Fixed: Multiple lookups from a single user before previous lookup is complete.
+
+### 🌟 Added
+- New: Autocomplete for server names.
+- New: DF S2 M+ scores.
+- New: M+ Season Cutoff data cached to speed up responses.
+- New: Aberrus raid added to raid progression.
+
+### 🗑️ Removed
+- Removed: SL S1 & S2 from M+ score output.
+
+## 1.2.0
+
+### 🛠️ Fixes
+- Fixed: Issue where no other button is pressed and user immediately selects the 'Done' button.
+
+---
+
+## 1.1.1
 
 ### 🪄 Updates
 
 - Updated: Command category to align with new naming convention.
 
+---
+
 ## 1.1.0
 
 ### 🛠️ Fixes
+
 - Fixed: Working again with the API changes made for DF.
 
 ### 🪄 Updates
@@ -27,4 +52,5 @@
 ## 1.0.0
 
 ### Added
+
 - Plugin created.